refactor: move setings to folder
