+++ frontend/src/game/index.ts
@@ -1,8 +1,8 @@
 import Character from './character';
-import Controls from './controls';
+import Controls from './services/controls';
 import CanvasHelper from './helperTypes/canvasHelper';
-import IControlsSettings from './controls/iControlsSettings.interface';
-import IGameSettings from './settings.interface';
+import IControlsSettings from './services/controls/iControlsSettings';
+import IGameSettings from './services/settings/iGameSettings';
 import LevelId from './levels/typeLevelIds';
 import Level from './levels';
 import levelList from './levels/list';
