+++ frontend/src/game/services/settings/iGameSettings.ts
